adds stack navigator
