+++ package.json
@@ -8,11 +8,16 @@
     "eject": "expo eject"
   },
   "dependencies": {
+    "@react-native-community/masked-view": "0.1.10",
+    "@react-navigation/native": "^5.9.0",
+    "@react-navigation/stack": "^5.13.0",
     "expo": "~40.0.0",
     "expo-status-bar": "~1.0.3",
     "react": "16.13.1",
     "react-dom": "16.13.1",
     "react-native": "https://github.com/expo/react-native/archive/sdk-40.0.1.tar.gz",
+    "react-native-reanimated": "2.0.0-rc.0",
+    "react-native-safe-area-context": "3.1.9",
     "react-native-svg": "12.1.0",
     "react-native-web": "~0.13.12"
   },
